storage: break dep of mz-storage-types on mz-persist-client
Pure refactor, no behavior change.

Accomplished by:
- moving `ShardId` and `PersistLocation` into mz-persist-types and
  leaving a `pub use` in mz-persist-client for convenience
- moving `PartStats` into mz-persist-types
- splitting a `PartStatsMetrics` out of `PushdownMetrics` and also
  moving that to mz-persist-types
