+++ src/persist-txn/src/lib.rs
@@ -211,10 +211,10 @@ use mz_dyncfg::{Config, ConfigSet};
 use mz_ore::instrument;
 use mz_persist_client::critical::SinceHandle;
 use mz_persist_client::error::UpperMismatch;
-use mz_persist_client::stats::PartStats;
 use mz_persist_client::write::WriteHandle;
-use mz_persist_client::{ShardId, ShardIdSchema};
-use mz_persist_types::codec_impls::VecU8Schema;
+use mz_persist_client::ShardId;
+use mz_persist_types::codec_impls::{ShardIdSchema, VecU8Schema};
+use mz_persist_types::stats::PartStats;
 use mz_persist_types::{Codec, Codec64, Opaque, StepForward};
 use serde::{Deserialize, Serialize};
 use timely::order::TotalOrder;
