+++ src/storage-types/build.rs
@@ -39,7 +39,6 @@ fn main() {
         .extern_path(".mz_repr.antichain", "::mz_repr::antichain")
         .extern_path(".mz_repr.global_id", "::mz_repr::global_id")
         .extern_path(".mz_orchestrator", "::mz_orchestrator")
-        .extern_path(".mz_persist_client", "::mz_persist_client")
         .extern_path(".mz_pgcopy.copy", "::mz_pgcopy")
         .extern_path(".mz_proto", "::mz_proto")
         .extern_path(".mz_repr.relation_and_scalar", "::mz_repr")
